requirements.txt
yfinance
mplfinance
pillow
torch
torchvision
setuptools
newsapi-python 
transformers 
torch
scikit-learn
numpy
wandb
pandas
gradio
seaborn
plotly
aiohttp>=3.8.0
asyncio
requests>=2.28.0